Updated demos to latest version of Touché framework.

+++ trunk/demos/TFCoreGraphicsDemo/ToucheFramework.framework/Versions/A/Headers/TFBlobPoint.h
@@ -33,9 +33,9 @@
 @property (assign) float y;
 
 + (id)point;
-+ (id)pointWithX:(float)xPos Y:(float)yPos;
++ (id)pointWithX:(CGFloat)xPos Y:(CGFloat)yPos;
 
-- (id)initWithX:(float)xPos Y:(float)yPos;
+- (id)initWithX:(CGFloat)xPos Y:(CGFloat)yPos;
 
 - (float)vectorLength;
 - (float)distanceFromBlobPoint:(TFBlobPoint*)otherPoint;
